refactor: extract util functions from auth.ts to utils.ts
