+++ src/auth/ui/statusBarItem.ts
@@ -10,10 +10,11 @@ import * as vscode from 'vscode'
 import { AwsContext } from '../../shared/awsContext'
 import { getIdeProperties } from '../../shared/extensionUtilities'
 import { DevSettings } from '../../shared/settings'
-import { Auth, login } from '../auth'
+import { Auth } from '../auth'
 import { getAllConnectionsInUse, onDidChangeConnections } from '../secondaryAuth'
 import { codicon, getIcon } from '../../shared/icons'
 import { throttle } from '../../shared/utilities/functionUtils'
+import { login } from '../utils'
 
 const statusbarPriority = 1
