+++ src/auth/ui/vue/show.ts
@@ -16,7 +16,7 @@ import {
     StaticProfile,
     StaticProfileKeyErrorMessage,
 } from '../../credentials/types'
-import { Auth, signout, tryAddCredentials } from '../../auth'
+import { Auth } from '../../auth'
 import { getCredentialFormatError, getCredentialsErrors } from '../../credentials/validation'
 import { profileExists } from '../../credentials/sharedCredentials'
 import { getLogger } from '../../../shared/logger'
@@ -26,6 +26,7 @@ import { CodeCatalystAuthenticationProvider } from '../../../codecatalyst/auth'
 import { getStartedCommand } from '../../../codecatalyst/explorer'
 import { ToolkitError } from '../../../shared/errors'
 import { isBuilderIdConnection } from '../../connection'
+import { tryAddCredentials, signout } from '../../utils'
 
 export class AuthWebview extends VueWebview {
     public override id: string = 'authWebview'
