+++ src/auth/utils.ts
@@ -3,9 +3,45 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { CommandDeclarations, Commands } from '../shared/vscode/commands2'
+import globals from '../shared/extensionGlobals'
+
+import * as nls from 'vscode-nls'
+const localize = nls.loadMessageBundle()
+
 import * as vscode from 'vscode'
+import * as localizedText from '../shared/localizedText'
+import { codicon, getIcon } from '../shared/icons'
+import { createQuickPick, DataQuickPickItem, showQuickPick } from '../shared/ui/pickerPrompter'
+import { isValidResponse } from '../shared/wizards/wizard'
+import { CancellationError } from '../shared/utilities/timeoutUtils'
+import { formatError, ToolkitError, UnknownError } from '../shared/errors'
+import { asString } from './providers/credentials'
+import { getResourceFromTreeNode } from '../shared/treeview/utils'
+import { Instance } from '../shared/utilities/typeConstructors'
+import { TreeNode } from '../shared/treeview/resourceTreeDataProvider'
+import { createInputBox } from '../shared/ui/inputPrompter'
+import { telemetry } from '../shared/telemetry/telemetry'
+import { createCommonButtons, createExitButton, createHelpButton, createRefreshButton } from '../shared/ui/buttons'
+import { getIdeProperties, isCloud9 } from '../shared/extensionUtilities'
+import { authHelpUrl } from '../shared/constants'
+import { getDependentAuths } from './secondaryAuth'
+import { DevSettings } from '../shared/settings'
+import { createRegionPrompter } from '../shared/ui/common/region'
+import { saveProfileToCredentials } from './credentials/sharedCredentials'
+import { SectionName, StaticProfile } from './credentials/types'
+import { throwOnInvalidCredentials } from './credentials/validation'
+import {
+    Connection,
+    createBuilderIdProfile,
+    createSsoProfile,
+    defaultSsoRegion,
+    hasScopes,
+    isBuilderIdConnection,
+    isSsoConnection,
+} from './connection'
+import { CommandDeclarations, Commands } from '../shared/vscode/commands2'
 import { showAuthWebview } from './ui/vue/show'
+import { Auth } from './auth'
 
 /**
  * The methods with backend logic for the Auth commands.
@@ -27,4 +63,515 @@ export class AuthCommandDeclarations implements CommandDeclarations<AuthCommandB
             id: 'aws.auth.showConnectionsPage',
         }),
     } as const
+}
+
+export async function promptForConnection(auth: Auth, type?: 'iam' | 'sso') {
+    const resp = await createConnectionPrompter(auth, type).prompt()
+    if (!isValidResponse(resp)) {
+        throw new CancellationError('user')
+    }
+
+    if (resp === 'addNewConnection') {
+        return addConnection.execute()
+    }
+
+    if (resp === 'editCredentials') {
+        return globals.awsContextCommands.onCommandEditCredentials()
+    }
+
+    return resp
+}
+
+export async function promptAndUseConnection(...[auth, type]: Parameters<typeof promptForConnection>) {
+    return telemetry.aws_setCredentials.run(async span => {
+        const resp = await promptForConnection(auth, type)
+        if (!resp) {
+            throw new CancellationError('user')
+        }
+
+        await auth.useConnection(resp)
+    })
+}
+
+const switchConnections = Commands.register('aws.auth.switchConnections', (auth: Auth | unknown) => {
+    telemetry.ui_click.emit({ elementId: 'devtools_connectToAws' })
+
+    if (auth instanceof Auth) {
+        return promptAndUseConnection(auth)
+    } else {
+        return promptAndUseConnection(getResourceFromTreeNode(auth, Instance(Auth)))
+    }
+})
+
+export async function signout(auth: Auth, conn: Connection | undefined = auth.activeConnection) {
+    if (conn?.type === 'sso') {
+        // TODO: does deleting the connection make sense UX-wise?
+        // this makes it disappear from the list of available connections
+        await auth.deleteConnection(conn)
+
+        const iamConnections = (await auth.listConnections()).filter(c => c.type === 'iam')
+        const fallbackConn = iamConnections.find(c => c.id === 'profile:default') ?? iamConnections[0]
+        if (fallbackConn !== undefined) {
+            await auth.useConnection(fallbackConn)
+        }
+    } else {
+        await auth.logout()
+
+        const fallbackConn = (await auth.listConnections()).find(c => c.type === 'sso')
+        if (fallbackConn !== undefined) {
+            await auth.useConnection(fallbackConn)
+        }
+    }
+}
+
+export const createBuilderIdItem = () =>
+    ({
+        label: codicon`${getIcon('vscode-person')} ${localize(
+            'aws.auth.builderIdItem.label',
+            'Use a personal email to sign up and sign in with {0}',
+            localizedText.builderId()
+        )}`,
+        data: 'builderId',
+        onClick: () => telemetry.ui_click.emit({ elementId: 'connection_optionBuilderID' }),
+        detail: `${localizedText.builderId()} is a new, personal profile for builders.`, // TODO: need a "Learn more" button ?
+    } as DataQuickPickItem<'builderId'>)
+
+export const createSsoItem = () =>
+    ({
+        label: codicon`${getIcon('vscode-organization')} ${localize(
+            'aws.auth.ssoItem.label',
+            'Connect using {0} {1}',
+            getIdeProperties().company,
+            localizedText.iamIdentityCenter
+        )}`,
+        data: 'sso',
+        onClick: () => telemetry.ui_click.emit({ elementId: 'connection_optionSSO' }),
+        detail: `Sign in to your company's ${localizedText.iamIdentityCenter} access portal login page.`,
+    } as DataQuickPickItem<'sso'>)
+
+export const createIamItem = () =>
+    ({
+        label: codicon`${getIcon('vscode-key')} ${localize('aws.auth.iamItem.label', 'Use IAM Credentials')}`,
+        data: 'iam',
+        onClick: () => telemetry.ui_click.emit({ elementId: 'connection_optionIAM' }),
+        detail: 'Activates working with resources in the Explorer. Not supported by CodeWhisperer. Requires an access key ID and secret access key.',
+    } as DataQuickPickItem<'iam'>)
+
+export async function createStartUrlPrompter(title: string, requiredScopes?: string[]) {
+    const existingConnections = (await Auth.instance.listConnections()).filter(isSsoConnection)
+
+    function validateSsoUrl(url: string) {
+        if (!url.match(/^(http|https):\/\//i)) {
+            return 'URLs must start with http:// or https://. Example: https://d-xxxxxxxxxx.awsapps.com/start'
+        }
+
+        try {
+            const uri = vscode.Uri.parse(url, true)
+            const isSameAuthority = (a: vscode.Uri, b: vscode.Uri) =>
+                a.authority.toLowerCase() === b.authority.toLowerCase()
+            const oldConn = existingConnections.find(conn => isSameAuthority(vscode.Uri.parse(conn.startUrl), uri))
+
+            if (oldConn && (!requiredScopes || hasScopes(oldConn, requiredScopes))) {
+                return 'A connection for this start URL already exists. Sign out before creating a new one.'
+            }
+        } catch (err) {
+            return `URL is malformed: ${UnknownError.cast(err).message}`
+        }
+    }
+
+    return createInputBox({
+        title: `${title}: Enter Start URL`,
+        placeholder: `Enter start URL for your organization's IAM Identity Center`,
+        buttons: [createHelpButton(), createExitButton()],
+        validateInput: validateSsoUrl,
+    })
+}
+
+export async function createBuilderIdConnection(auth: Auth, scopes?: string[]) {
+    const newProfile = createBuilderIdProfile(scopes)
+    const existingConn = (await auth.listConnections()).find(isBuilderIdConnection)
+    if (!existingConn) {
+        return auth.createConnection(newProfile)
+    }
+
+    const userResponse = await promptLogoutExistingBuilderIdConnection()
+    if (userResponse !== 'signout') {
+        throw new CancellationError('user')
+    }
+
+    await signout(auth, existingConn)
+
+    return auth.createConnection(newProfile)
+}
+
+/**
+ * Prompts the user to log out of an existing Builder ID connection.
+ *
+ * @returns The name of the action performed by the user
+ */
+async function promptLogoutExistingBuilderIdConnection(): Promise<'signout' | 'cancel'> {
+    const items: DataQuickPickItem<'signout' | 'cancel'>[] = [
+        {
+            data: 'signout',
+            label: `Currently signed in with ${getIdeProperties().company} Builder ID. Sign out to add another?`,
+            detail: `This will sign out of your current ${
+                getIdeProperties().company
+            } Builder ID and open the sign-in page in browser.`,
+        },
+        { data: 'cancel', label: 'Cancel' },
+    ]
+    const resp = await showQuickPick(items, {
+        title: `Sign in to different ${getIdeProperties().company} Builder ID`,
+        buttons: createCommonButtons() as vscode.QuickInputButton[],
+    })
+
+    return resp === undefined ? 'cancel' : resp
+}
+
+export async function showRegionPrompter(
+    title: string = `IAM Identity Center: Select Region`,
+    placeholder: string = `Select region for your organization's IAM Identity Center`
+) {
+    const region = await createRegionPrompter(undefined, {
+        defaultRegion: defaultSsoRegion,
+        buttons: createCommonButtons(),
+        title: title,
+        placeholder: placeholder,
+    }).prompt()
+
+    if (!isValidResponse(region)) {
+        throw new CancellationError('user')
+    }
+    telemetry.ui_click.emit({ elementId: 'connection_region' })
+
+    return region
+}
+
+Commands.register('aws.auth.help', async () => {
+    vscode.env.openExternal(vscode.Uri.parse(authHelpUrl))
+    telemetry.aws_help.emit()
+})
+
+Commands.register('aws.auth.signout', () => {
+    telemetry.ui_click.emit({ elementId: 'devtools_signout' })
+    return signout(Auth.instance)
+})
+
+const addConnection = Commands.register({ id: 'aws.auth.addConnection', telemetryThrottleMs: false }, async () => {
+    const c9IamItem = createIamItem()
+    c9IamItem.detail =
+        'Activates working with resources in the Explorer. Requires an access key ID and secret access key.'
+    const items = isCloud9() ? [createSsoItem(), c9IamItem] : [createBuilderIdItem(), createSsoItem(), createIamItem()]
+
+    const resp = await showQuickPick(items, {
+        title: localize('aws.auth.addConnection.title', 'Add a Connection to {0}', getIdeProperties().company),
+        placeholder: localize('aws.auth.addConnection.placeholder', 'Select a connection option'),
+        buttons: createCommonButtons() as vscode.QuickInputButton[],
+    })
+    if (!isValidResponse(resp)) {
+        telemetry.ui_click.emit({ elementId: 'connection_optionescapecancel' })
+        throw new CancellationError('user')
+    }
+
+    switch (resp) {
+        case 'iam':
+            return await globals.awsContextCommands.onCommandCreateCredentialsProfile()
+        case 'sso': {
+            const startUrlPrompter = await createStartUrlPrompter('IAM Identity Center')
+            const startUrl = await startUrlPrompter.prompt()
+            if (!isValidResponse(startUrl)) {
+                throw new CancellationError('user')
+            }
+            telemetry.ui_click.emit({ elementId: 'connection_startUrl' })
+
+            const region = await showRegionPrompter()
+
+            const conn = await Auth.instance.createConnection(createSsoProfile(startUrl, region.id))
+            return Auth.instance.useConnection(conn)
+        }
+        case 'builderId': {
+            return createBuilderIdConnection(Auth.instance)
+        }
+    }
+})
+
+export async function tryAddCredentials(
+    profileName: SectionName,
+    profileData: StaticProfile,
+    tryConnect = true
+): Promise<boolean> {
+    const auth = Auth.instance
+
+    // sanity checks
+    await throwOnInvalidCredentials(profileName, profileData)
+    const authenticationError = await auth.authenticateData(profileData)
+    if (authenticationError) {
+        throw new ToolkitError(`Found error with '${authenticationError.key}':'${authenticationError.error}' `, {
+            code: 'InvalidCredentials',
+        })
+    }
+
+    await saveProfileToCredentials(profileName, profileData)
+
+    if (tryConnect) {
+        const id = asString({
+            credentialSource: 'profile',
+            credentialTypeId: profileName,
+        })
+        const conn = await auth.getConnection({ id })
+
+        if (conn === undefined) {
+            throw new ToolkitError(`Failed to get connection from profile: ${profileName}`, {
+                code: 'MissingConnection',
+            })
+        }
+
+        await auth.useConnection(conn)
+    }
+    return true
+}
+
+const getConnectionIcon = (conn: Connection) =>
+    conn.type === 'sso' ? getIcon('vscode-account') : getIcon('vscode-key')
+
+export function createConnectionPrompter(auth: Auth, type?: 'iam' | 'sso') {
+    const addNewConnection = {
+        label: codicon`${getIcon('vscode-plus')} Add New Connection`,
+        data: 'addNewConnection' as const,
+    }
+    const editCredentials = {
+        label: codicon`${getIcon('vscode-pencil')} Edit Credentials`,
+        data: 'editCredentials' as const,
+    }
+    const placeholder =
+        type === 'iam'
+            ? localize('aws.auth.promptConnection.iam.placeholder', 'Select an IAM credential')
+            : localize('aws.auth.promptConnection.all.placeholder', 'Select a connection')
+
+    const refreshButton = createRefreshButton()
+    refreshButton.onClick = () => void prompter.clearAndLoadItems(loadItems())
+
+    // Place add/edit connection items at the bottom, then sort 'sso' connections
+    // first, then valid connections, then finally the item label
+    function getSortOrder(item: DataQuickPickItem<Connection | string>) {
+        if (item.data === addNewConnection.data) {
+            return 10
+        } else if (item.data === editCredentials.data) {
+            return 9
+        }
+
+        const conn = item.data as Connection
+        if (conn.type === 'sso') {
+            return 0
+        } else if (auth.getConnectionState(conn) === 'valid') {
+            return 1
+        }
+
+        return 2
+    }
+
+    const prompter = createQuickPick(loadItems(), {
+        placeholder,
+        title: localize('aws.auth.promptConnection.title', 'Switch Connection'),
+        buttons: [refreshButton, createExitButton()],
+        compare: (a, b) => {
+            if (getSortOrder(a) === 0 && getSortOrder(b) === 0) {
+                return a.label.localeCompare(b.label)
+            }
+
+            return getSortOrder(a) - getSortOrder(b)
+        },
+    })
+
+    return prompter
+
+    async function* loadItems(): AsyncGenerator<
+        DataQuickPickItem<Connection | 'addNewConnection' | 'editCredentials'>[]
+    > {
+        const connections = auth.listAndTraverseConnections()
+
+        let hasShownEdit = false
+
+        yield [addNewConnection]
+        for await (const conn of connections) {
+            if (conn.label.includes('profile:') && !hasShownEdit) {
+                hasShownEdit = true
+                yield [toPickerItem(conn), editCredentials]
+            } else {
+                yield [toPickerItem(conn)]
+            }
+        }
+    }
+
+    function toPickerItem(conn: Connection): DataQuickPickItem<Connection> {
+        const state = auth.getConnectionState(conn)
+        if (state === 'valid') {
+            return {
+                data: conn,
+                label: codicon`${getConnectionIcon(conn)} ${conn.label}`,
+                description: getConnectionDescription(conn),
+            }
+        }
+
+        const getDetail = () => {
+            if (!DevSettings.instance.get('renderDebugDetails', false)) {
+                return
+            }
+
+            const err = auth.getInvalidationReason(conn)
+            return err ? formatError(err) : undefined
+        }
+
+        return {
+            detail: getDetail(),
+            data: conn,
+            invalidSelection: state !== 'authenticating',
+            label: codicon`${getIcon('vscode-error')} ${conn.label}`,
+            description:
+                state === 'authenticating'
+                    ? 'authenticating...'
+                    : localize(
+                          'aws.auth.promptConnection.expired.description',
+                          'Expired or Invalid, select to authenticate'
+                      ),
+            onClick:
+                state !== 'authenticating'
+                    ? async () => {
+                          // XXX: this is hack because only 1 picker can be shown at a time
+                          // Some legacy auth providers will show a picker, hiding this one
+                          // If we detect this then we'll jump straight into using the connection
+                          let hidden = false
+                          const sub = prompter.quickPick.onDidHide(() => {
+                              hidden = true
+                              sub.dispose()
+                          })
+                          const newConn = await reauthCommand.execute(auth, conn)
+                          if (hidden && newConn && auth.getConnectionState(newConn) === 'valid') {
+                              await auth.useConnection(newConn)
+                          } else {
+                              await prompter.clearAndLoadItems(loadItems())
+                              prompter.selectItems(
+                                  ...prompter.quickPick.items.filter(i => i.label.includes(conn.label))
+                              )
+                          }
+                      }
+                    : undefined,
+        }
+    }
+
+    function getConnectionDescription(conn: Connection) {
+        if (conn.type === 'iam') {
+            // TODO: implement a proper `getConnectionSource` method to discover where a connection came from
+            const descSuffix = conn.id.startsWith('profile:')
+                ? 'configured locally (~/.aws/config)'
+                : conn.id.startsWith('sso:')
+                ? 'sourced from IAM Identity Center'
+                : 'sourced from the environment'
+
+            return `IAM Credential, ${descSuffix}`
+        }
+
+        const toolAuths = getDependentAuths(conn)
+        if (toolAuths.length === 0) {
+            return undefined
+        } else if (toolAuths.length === 1) {
+            return `Connected to ${toolAuths[0].toolLabel}`
+        } else {
+            return `Connected to Dev Tools`
+        }
+    }
+}
+
+export const reauthCommand = Commands.register('_aws.auth.reauthenticate', async (auth: Auth, conn: Connection) => {
+    try {
+        return await auth.reauthenticate(conn)
+    } catch (err) {
+        throw ToolkitError.chain(err, 'Unable to authenticate connection')
+    }
+})
+
+// Used to decouple from the `Commands` implementation
+Commands.register('_aws.auth.autoConnect', () => Auth.instance.tryAutoConnect())
+
+export const useIamCredentials = Commands.register('_aws.auth.useIamCredentials', (auth: Auth) => {
+    telemetry.ui_click.emit({ elementId: 'explorer_IAMselect_VSCode' })
+
+    return promptAndUseConnection(auth, 'iam')
+})
+
+// Legacy commands
+export const login = Commands.register('aws.login', async (auth: Auth = Auth.instance) => {
+    const connections = await auth.listConnections()
+    if (connections.length === 0) {
+        return addConnection.execute()
+    } else {
+        return switchConnections.execute(auth)
+    }
+})
+Commands.register('aws.logout', () => signout(Auth.instance))
+Commands.register('aws.credentials.edit', () => globals.awsContextCommands.onCommandEditCredentials())
+Commands.register('aws.credentials.profile.create', async () => {
+    try {
+        await globals.awsContextCommands.onCommandCreateCredentialsProfile()
+    } finally {
+        telemetry.aws_createCredentials.emit()
+    }
+})
+
+function mapEventType<T, U = void>(event: vscode.Event<T>, fn?: (val: T) => U): vscode.Event<U> {
+    const emitter = new vscode.EventEmitter<U>()
+    event(val => (fn ? emitter.fire(fn(val)) : emitter.fire(undefined as U)))
+
+    return emitter.event
+}
+
+export class AuthNode implements TreeNode<Auth> {
+    public readonly id = 'auth'
+    public readonly onDidChangeTreeItem = mapEventType(this.resource.onDidChangeActiveConnection)
+
+    public constructor(public readonly resource: Auth) {}
+
+    public getTreeItem() {
+        // Calling this here is robust but `TreeShim` must be instantiated lazily to stop side-effects
+        this.resource.tryAutoConnect()
+
+        if (!this.resource.hasConnections) {
+            const item = new vscode.TreeItem(`Connect to ${getIdeProperties().company} to Get Started...`)
+            item.command = addConnection.build().asCommand({ title: 'Add Connection' })
+
+            return item
+        }
+
+        const conn = this.resource.activeConnection
+        const itemLabel =
+            conn?.label !== undefined
+                ? localize('aws.auth.node.connected', `Connected with {0}`, conn.label)
+                : localize('aws.auth.node.selectConnection', 'Select a connection...')
+
+        const item = new vscode.TreeItem(itemLabel)
+        item.contextValue = 'awsAuthNode'
+
+        if (conn !== undefined && conn.state !== 'valid') {
+            item.iconPath = getIcon('vscode-error')
+            if (conn.state === 'authenticating') {
+                this.setDescription(item, 'authenticating...')
+            } else {
+                this.setDescription(item, 'expired or invalid, click to authenticate')
+                item.command = reauthCommand.build(this.resource, conn).asCommand({ title: 'Reauthenticate' })
+            }
+        } else {
+            item.command = switchConnections.build(this.resource).asCommand({ title: 'Login' })
+            item.iconPath = conn !== undefined ? getConnectionIcon(conn) : undefined
+        }
+
+        return item
+    }
+
+    private setDescription(item: vscode.TreeItem, text: string) {
+        if (isCloud9()) {
+            item.tooltip = item.tooltip ?? text
+        } else {
+            item.description = text
+        }
+    }
 }
